Add failure domain selection
Add the Node view to complement the unused Node model.  Add
interactive selection support to the Node, Zone & Region groups in
addition to the Device.

To better support rings with many devices the device/weight/bars have
been moved to the bottom and the ring visualization was moved to a
seperate html element with a fixed posistion so that you can scroll
down to select Nodes/Devices and still see the visualization.